Extension refactoring
