--- Packages/CBSL/Core/Runtime/Extensions/Collider.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: 8435e55e28ec4378a4aa63850d5db615
-folderAsset: yes
-DefaultImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
