+++ Packages/CBSL/Core/Runtime/Extensions/Collider2DExtensions.cs
@@ -1,10 +1,10 @@
 ﻿using UnityEngine;
 
-namespace CBSL.Core.Extensions.Collider2D {
+namespace CBSL.Core.Extensions {
 
-    public static class Grounded2D {
+    public static class Collider2DExtensions {
 
-        public static bool IsGrounded(this BoxCollider2D collider, LayerMask layerMask, Vector2 direction, float distance = 0.1f, float angle = 0f) {
+        public static bool IsGrounded(this Collider2D collider, LayerMask layerMask, Vector2 direction, float distance = 0.1f, float angle = 0f) {
             var bounds = collider.bounds;
             
             var result = Physics2D.BoxCast(bounds.center, bounds.size, angle,
