--- Packages/CBSL/Core/Runtime/Extensions/Transform.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: be707065ac7fc4fd5a4aa1deedf1fce1
-folderAsset: yes
-DefaultImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
